Remove user and group from configuration
It is not used, we are using a TCP socket, so the ownership of the UNIX
socket does not apply. Remove because when running the container with a
user other than root php-fpm generates a warning that these
configuration do not apply in that case. This gets rid of the warning
and since we are not using these configurations we don't need them.